Refactored packaging to mvc structure
